Add route to get the correction comment of one student.
